pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "livedoc"
version = "0.1.0"
description = "Document processing pipeline for converting PDFs to structured reports using vision LLMs"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "LiveDoc Team"}
]
keywords = ["pdf", "llm", "document-processing", "vision", "report-generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "ollama",
    "pdf2image",
    "Pillow",
    "PyYAML",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "mypy",
    "ruff",
]

[project.scripts]
livedoc = "livedoc.cli:main"

[project.urls]
Homepage = "https://github.com/example/livedoc"
Documentation = "https://github.com/example/livedoc#readme"
Repository = "https://github.com/example/livedoc"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
livedoc = ["py.typed"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
